+++ packages/tailwindcss/src/utilities.test.ts
@@ -147,7 +147,6 @@ test('inset', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --spacing-4: 1rem;
-      --inset-shadow-sm: inset 0 1px 1px #0000000d;
       --inset-shadowned: 1940px;
     }
 
@@ -3162,10 +3161,6 @@ describe('container', () => {
     ).toMatchInlineSnapshot(`
       ":root, :host {
         --breakpoint-sm: 40rem;
-        --breakpoint-md: 48rem;
-        --breakpoint-lg: 64rem;
-        --breakpoint-xl: 80rem;
-        --breakpoint-2xl: 96rem;
       }
 
       .container {
@@ -3305,10 +3300,6 @@ describe('container', () => {
     ).toMatchInlineSnapshot(`
       ":root, :host {
         --breakpoint-sm: 40rem;
-        --breakpoint-md: 48rem;
-        --breakpoint-lg: 64rem;
-        --breakpoint-xl: 80rem;
-        --breakpoint-2xl: 96rem;
       }
 
       .container {
@@ -4045,11 +4036,7 @@ test('translate-x', async () => {
       ['translate-x-full', '-translate-x-full', 'translate-x-px', '-translate-x-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --spacing: .25rem;
-    }
-
-    .-translate-x-\\[var\\(--value\\)\\] {
+    ".-translate-x-\\[var\\(--value\\)\\] {
       --tw-translate-x: calc(var(--value) * -1);
       translate: var(--tw-translate-x) var(--tw-translate-y);
     }
@@ -4173,11 +4160,7 @@ test('translate-y', async () => {
       ['translate-y-full', '-translate-y-full', 'translate-y-px', '-translate-y-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --spacing: .25rem;
-    }
-
-    .-translate-y-\\[var\\(--value\\)\\] {
+    ".-translate-y-\\[var\\(--value\\)\\] {
       --tw-translate-y: calc(var(--value) * -1);
       translate: var(--tw-translate-x) var(--tw-translate-y);
     }
@@ -7734,11 +7717,7 @@ test('divide-x with custom default border width', async () => {
       ['divide-x'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --default-border-width: 2px;
-    }
-
-    :where(.divide-x > :not(:last-child)) {
+    ":where(.divide-x > :not(:last-child)) {
       --tw-divide-x-reverse: 0;
       border-inline-style: var(--tw-border-style);
       border-inline-start-width: calc(2px * var(--tw-divide-x-reverse));
@@ -7840,11 +7819,7 @@ test('divide-y with custom default border width', async () => {
       ['divide-y'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --default-border-width: 2px;
-    }
-
-    :where(.divide-y > :not(:last-child)) {
+    ":where(.divide-y > :not(:last-child)) {
       --tw-divide-y-reverse: 0;
       border-bottom-style: var(--tw-border-style);
       border-top-style: var(--tw-border-style);
@@ -9930,11 +9905,7 @@ test('border with custom default border width', async () => {
       ['border'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --default-border-width: 2px;
-    }
-
-    .border {
+    ".border {
       border-style: var(--tw-border-style);
       border-width: 2px;
     }
@@ -13870,7 +13841,6 @@ test('transition', async () => {
     ":root, :host {
       --default-transition-timing-function: ease;
       --default-transition-duration: .1s;
-      --transition-property: color, background-color, border-color, text-decoration-color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter;
       --transition-property-opacity: opacity;
     }
 
@@ -13936,12 +13906,7 @@ test('transition', async () => {
       ['transition', 'transition-all', 'transition-colors'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root, :host {
-      --default-transition-timing-function: ease;
-      --default-transition-duration: .1s;
-    }
-
-    .transition {
+    ".transition {
       transition-property: color, background-color, border-color, outline-color, text-decoration-color, fill, stroke, --tw-gradient-from, --tw-gradient-via, --tw-gradient-to, opacity, box-shadow, transform, translate, scale, rotate, filter, -webkit-backdrop-filter, backdrop-filter;
       transition-timing-function: var(--tw-ease, ease);
       transition-duration: var(--tw-duration, .1s);
@@ -15227,8 +15192,6 @@ test('shadow', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --color-red-500: #ef4444;
-      --shadow-sm: 0 1px 3px 0 #0000001a, 0 1px 2px -1px #0000001a;
-      --shadow-xl: 0 20px 25px -5px #0000001a, 0 8px 10px -6px #0000001a;
     }
 
     .shadow-\\[10px_10px\\] {
@@ -15451,8 +15414,6 @@ test('inset-shadow', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --color-red-500: #ef4444;
-      --inset-shadow: inset 0 2px 4px #0000000d;
-      --inset-shadow-sm: inset 0 1px 1px #0000000d;
     }
 
     .inset-shadow {
@@ -16443,7 +16404,6 @@ describe('spacing utilities', () => {
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       ":root, :host {
         --spacing-sm: 8px;
-        --container-sm: 256px;
       }
 
       .w-sm {
@@ -16496,6 +16456,40 @@ describe('custom utilities', () => {
             text-box-edge: cap alphabetic;
           }
         }
+      }"
+    `)
+  })
+
+  test('custom static utility emit CSS variables if the utility is used', async () => {
+    let { build } = await compile(css`
+      @layer utilities {
+        @tailwind utilities;
+      }
+
+      @theme {
+        --example-foo: 123px;
+      }
+
+      @utility foo {
+        value: var(--example-foo);
+      }
+    `)
+    let compiled = build([])
+
+    // `foo` is not used yet:
+    expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`"@layer utilities;"`)
+
+    // `foo` is used, and the CSS variable is emitted:
+    compiled = build(['foo'])
+    expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .foo {
+          value: var(--example-foo);
+        }
+      }
+
+      :root, :host {
+        --example-foo: 123px;
       }"
     `)
   })
@@ -17650,6 +17644,56 @@ describe('custom utilities', () => {
       `)
       expect(await compileCss(input, ['example-foo', 'example-xs/foo'])).toEqual('')
     })
+
+    test('variables used in `@utility` will not be emitted if the utility is not used', async () => {
+      let input = css`
+        @theme {
+          --example-foo: red;
+          --color-red-500: #f00;
+        }
+
+        @utility example-* {
+          color: var(--color-red-500);
+          background-color: --value(--example);
+        }
+
+        @tailwind utilities;
+      `
+
+      expect(await compileCss(input, ['flex'])).toMatchInlineSnapshot(`
+        ".flex {
+          display: flex;
+        }"
+      `)
+    })
+
+    test('variables used in `@utility` will be emitted if the utility is used', async () => {
+      let input = css`
+        @theme {
+          --example-foo: red;
+          --color-red-500: #f00;
+        }
+
+        @utility example-* {
+          color: var(--color-red-500);
+          background-color: --value(--example);
+        }
+
+        @tailwind utilities;
+      `
+
+      expect(await compileCss(input, ['example-foo'])).toMatchInlineSnapshot(`
+        ":root, :host {
+          --example-foo: red;
+          --color-red-500: red;
+        }
+
+        .example-foo {
+          color: var(--color-red-500);
+          background-color: var(--example-foo);
+        }"
+      `)
+    })
   })
 
   test('resolve value based on `@theme`', async () => {
@@ -17710,11 +17754,7 @@ describe('custom utilities', () => {
     `
 
     expect(await compileCss(input, ['tab-github'])).toMatchInlineSnapshot(`
-      ":root, :host {
-        --tab-size-github: 8;
-      }
-
-      .tab-github {
+      ".tab-github {
         tab-size: 8;
       }"
     `)