+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Only expose used CSS variables ([#16211](https://github.com/tailwindlabs/tailwindcss/pull/16211))
+- Only expose used `@keyframes` ([#16211](https://github.com/tailwindlabs/tailwindcss/pull/16211))
 
 ## [4.0.4] - 2025-02-06